+++ src/main/java/com/example/demo/jukebox/dao/SongDao.java
@@ -8,8 +8,8 @@ import com.example.demo.jukebox.dto.Song;
 
 @Mapper
 public interface SongDao {
-  public Long insert(Song song);
+  public int insert(Song song);
   public Song selectBySid(Long sid);
   public List<Song> selectByJid(Long jid);
-  public Long delete(Long sid);
+  public int delete(Long sid);
 }
